requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.5
djangorestframework==3.15.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langgraph==0.2.68
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.2
msgpack==1.1.0
multidict==6.1.0
numpy==2.2.2
openai==1.60.2
orjson==3.10.15
packaging==24.2
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0
